Rename and improve PriorityList to PrioritySet
